update functionalities no parameter

+++ src/__init__.py
@@ -29,19 +29,28 @@ def words(word_id):
             resp = [{"id": word.id, "word": word.name} for word in Word.query.all()]
     elif request.method == "POST":
         data = request.get_json()
-        word = Word(name=data['word'])
-        db.session.add(word)
-        db.session.commit()
-        status, resp = 201, {"id": word.id, "message": "Word added successfully"}
+        if data.get("word") != None:
+            word = Word(name=data['word'])
+            db.session.add(word)
+            db.session.commit()
+            status, resp = 201, {"id": word.id, "message": "Word added successfully"}
+        else:
+            status, resp = 400, {"message": "missing parameter"}
     elif request.method == "PUT":
         data = request.get_json()
-        word = Word.query.filter_by(id=data['id']).first()
-        word.name = data['word']
-        db.session.commit()
-        resp = {"id": word.id, "message": "Word updated successfully"}
+        if data.get("id") != None and  data.get("word") != None:
+            word = Word.query.filter_by(id=data['id']).first()
+            word.name = data['word']
+            db.session.commit()
+            resp = {"id": word.id, "message": "Word updated successfully"}
+        else:
+            status, resp = 400, {"message": "missing parameter"}
     elif request.method == "DELETE":
         data = request.get_json()
-        _id = Word.query.filter(Word.name==data['word']).delete()
-        resp = {"id": _id, "message": "Word deleted successfully"}
+        if data.get("word") != None:
+            _id = Word.query.filter(Word.name==data['word']).delete()
+            resp = {"id": _id, "message": "Word deleted successfully"}
+        else:
+            status, resp = 400, {"message": "missing parameter"}
     return jsonify(resp), status
